+++ src/classes/tasks/edit.php
@@ -7,7 +7,6 @@ class Edit
     public static function run(array $params)
     {
 
-        $aeDebug=\AeSecure\Debug::getInstance();
         $aeSettings=\AeSecure\Settings::getInstance();
 
         header("Cache-Control: no-store, no-cache, must-revalidate, max-age=0");