+++ src/router.php
@@ -8,13 +8,28 @@ $filename=\AeSecure\Functions::getParam('file', 'string', '', false);
 // Check the optional format parameter.  If equal to 'slides', the task will be 'slideshow', 'display' otherwise
 $format=\AeSecure\Functions::getParam('format', 'string', 'html', false, 8);
 
-if (!(in_array($format, array('html','slides')))) $format='html';
+// Only these format are recognized.  Default : html
+if (!(in_array($format, array('htm','html','pdf','slides')))) $format='html';
 
 if ($filename!=='') {
 
     require_once __DIR__.'/classes/markdown.php';
 
-    $task=($format==='slides') ? 'slideshow' : 'display';
+    switch ($format) {
+
+        case 'pdf':
+            $task='pdf';
+            break;
+
+        case 'slides':
+            $task='slideshow';
+            break;
+
+        default:                  // htm or html
+            $task='display';
+            break;
+    } // switch
+
 
     $filename = str_replace('/', DIRECTORY_SEPARATOR, str_replace('docs/', '', \AeSecure\Files::removeExtension($filename))).'.md';